studying/views.py
from rest_framework import viewsets

from studying.models import Course
from studying.serializers import CourseSerializer


class CourseViewSet(viewsets.ModelViewSet):
    serializer_class = CourseSerializer
    queryset = Course.objects.all()
